requirements.txt
asgiref==3.8.1
certifi==2025.1.31
charset-normalizer==3.4.1
Django==5.2
djangorestframework==3.16.0
idna==3.10
pillow==11.2.1
requests==2.32.3
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.4.0
